.claude/skills/process/scripts/requirements.txt
youtube-transcript-api==1.2.3
cloudscraper
beautifulsoup4
google-genai
python-dotenv
anthropic
